# backend/docs/swagger.yaml
basePath: /
definitions:
...
  backend_internal_domain.Email:
    type: object
...
  backend_internal_domain.Permission:
    properties:
      action:
        type: string
      createdAt:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      resource:
        type: string
      updatedAt:
        type: string
    type: object
...
  backend_internal_domain.Role:
    properties:
      createdAt:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      permissions:
        items:
          $ref: '#/definitions/backend_internal_domain.Permission'
        type: array
      updatedAt:
        type: string
    type: object
...
  backend_internal_domain.User:
    properties:
      confirmedAt:
        type: string
      createdAt:
        type: string
      email:
        $ref: '#/definitions/backend_internal_domain.Email'
      id:
        type: integer
      role:
        $ref: '#/definitions/backend_internal_domain.Role'
      roleId:
        type: integer
      updatedAt:
        type: string
    type: object
...
  internal_adapters_http.updateProfileRequest:
    properties:
      currentPassword:
        type: string
      email:
        type: string
      password:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    name: Kthulu Framework
    url: https://github.com/kthulu-framework
  description: |-
    Complete enterprise-grade API for the Kthulu framework

    This API provides authentication, user management, organization management,
    and ERP-lite modules including contacts, products, invoices, inventory, and calendar.

    ## Authentication

    Most endpoints require authentication via JWT Bearer token in the Authorization header:
    ```
    Authorization: Bearer <your-jwt-token>
    ```

    ## Module Tags

    Endpoints are tagged by module for selective generation:
    - `@kthulu:core` - Core framework functionality
    - `@kthulu:module:auth` - Authentication module
    - `@kthulu:module:user` - User management module
    - `@kthulu:module:org` - Organization management module
    - `@kthulu:module:contacts` - Contacts module
    - `@kthulu:module:products` - Products module
    - `@kthulu:module:invoices` - Invoices module
    - `@kthulu:module:inventory` - Inventory module
    - `@kthulu:module:calendar` - Calendar module
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Kthulu API
  version: 1.0.0
paths:
...
  /users/me:
    get:
      description: Returns the authenticated user's profile information
      produces:
      - application/json
      responses:
        "200":
          description: User profile retrieved successfully
          schema:
            $ref: '#/definitions/backend_internal_domain.User'
        "401":
          description: Unauthorized - invalid or missing token
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: User not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get current user profile
      tags:
      - Users
    patch:
      consumes:
      - application/json
      description: Updates the authenticated user's profile information
      parameters:
      - description: Profile update details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_adapters_http.updateProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Profile updated successfully
          schema:
            $ref: '#/definitions/backend_internal_domain.User'
        "400":
          description: Invalid request or current password
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized - invalid or missing token
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: User not found
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Email already exists
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update current user profile
      tags:
      - Users
securityDefinitions:
  BearerAuth:
    description: 'JWT token obtained from /auth/login endpoint. Format: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"